requirements.txt
streamlit
tensorflow
opencv-python
mediapipe
gtts
pygame
Pillow
